Lay groundwork for AI integration
Summary of commit:
    - VectorInt3 renamed to IntVector3 for legibility.
    - GameManager moving from bools to enums for tracking state.
    - Temporary functions and variables while the Player and Menu systems are being developed

Player is an ABC interface that will be inherited by classes Human, EasyAI, MediumAI, etc. Its pure virtual function getMove(const CuboidArray& ca) will be called in the update() function GameManager, and the derived classes will each supply a valid move according to their own algorithms.

The Menu system consists of two classes, Menu and MenuItem. Menus possess an arbitrary number of MenuItems. MenuItems hold a string code and an address to a function which changes the game state. When input is received in GameManager, it is passed to the current active Menu along with an unassigned function pointer, which then passes both on to its MenuItems. If any of its MenuItems are holding a code which matches the input, that MenuItem will regurgitate the function pointer it was storing.

The Menu/MenuItem hierarchy is constructed at compile time, possibly in GameManager.